test for nextflow/rnafusion models
