+++ tests/cli/workflow/rnafusion/test_cli_rnafusion_compound_commands.py
@@ -7,7 +7,7 @@ from cg.cli.workflow.rnafusion.base import rnafusion, start, store, start_availa
 from cg.models.cg_config import CGConfig
 from click.testing import CliRunner
 from cg.meta.workflow.rnafusion import RnafusionAnalysisAPI
-from cg.constants import EXIT_SUCCESS, EXIT_FAIL
+from cg.constants import EXIT_SUCCESS
 
 
 def test_rnafusion_no_args(cli_runner: CliRunner, rnafusion_context: CGConfig):
