+++ tests/models/nextflow/conftest.py
@@ -1,51 +1,37 @@
 import pytest
 
 
-@pytest.fixture(name="rnafusion_sample")
-def fixture_rnafusion_sample() -> str:
-    """Return sample"""
-    return "rnafusion_sample"
+@pytest.fixture(name="nextflow_deliverables")
+def fixture_nextflow_deliverables() -> dict:
+    return {
+        "format": "tsv",
+        "id": "CASEID",
+        "path": "PATHTOCASE/fusioncatcher/CASEID.fusioncatcher.fusion-genes.txt",
+        "path_index": "~",
+        "step": "fusioncatcher",
+        "tag": "fusioncatcher",
+    }
 
-@pytest.fixture(name="rnafusion_fastq_r1")
-def fixture_rnafusion_fastq_r1() -> list:
-    """Return fastq_r1 list"""
-    return [
-        "dir/XXXXXXXX1_000000_S000_L001_R1_001.fastq.gz",
-        "dir/XXXXXXXX2_000000_S000_L001_R1_001.fastq.gz",
-        "dir/XXXXXXXX3_000000_S000_L001_R1_001.fastq.gz",
-        "dir/XXXXXXXX4_000000_S000_L001_R1_001.fastq.gz",
-    ]
 
-@pytest.fixture(name="rnafusion_fastq_r2_same_length")
-def fixture_rnafusion_fastq_r2_same_length() -> list:
-    """Return fastq_r2 list, same length as fastq_r1"""
-    return [
-        "dir/XXXXXXXX1_000000_S000_L001_R2_001.fastq.gz",
-        "dir/XXXXXXXX2_000000_S000_L001_R2_001.fastq.gz",
-        "dir/XXXXXXXX3_000000_S000_L001_R2_001.fastq.gz",
-        "dir/XXXXXXXX4_000000_S000_L001_R2_001.fastq.gz",
-    ]
+@pytest.fixture(name="nextflow_deliverables_with_empty_entry")
+def fixture_nextflow_deliverables_with_empty_entry() -> dict:
+    return {
+        "format": "",
+        "id": "CASEID",
+        "path": "PATHTOCASE/fusioncatcher/CASEID.fusioncatcher.fusion-genes.txt",
+        "path_index": "~",
+        "step": "fusioncatcher",
+        "tag": "fusioncatcher",
+    }
 
-@pytest.fixture(name="rnafusion_fastq_r2_not_same_length")
-def fixture_rnafusion_fastq_r2_not_same_length() -> list:
-    """Return fastq_r2 list, different length as fastq_r1"""
-    return [
-        "dir/XXXXXXXX1_000000_S000_L001_R2_001.fastq.gz",
-        "dir/XXXXXXXX2_000000_S000_L001_R2_001.fastq.gz",
-        "dir/XXXXXXXX3_000000_S000_L001_R2_001.fastq.gz",
-    ]
 
-@pytest.fixture(name="rnafusion_fastq_r2_empty")
-def fixture_rnafusion_fastq_r2_empty() -> list:
-    """Return empty fastq_r2 list"""
-    return []
-
-@pytest.fixture(name="rnafusion_strandedness_acceptable")
-def fixture_rnafusion_strandedness_acceptable() -> list:
-    """Return an acceptable strandedness"""
-    return "reverse"
-
-@pytest.fixture(name="rnafusion_strandedness_not_acceptable")
-def fixture_rnafusion_strandedness_not_acceptable() -> list:
-    """Return a not acceptable strandedness"""
-    return "double_stranded"
+@pytest.fixture(name="nextflow_deliverables_with_faulty_entry")
+def fixture_nextflow_deliverables_with_faulty_entry() -> dict:
+    return {
+        "faulty_header": "tsv",
+        "id": "CASEID",
+        "path": "PATHTOCASE/fusioncatcher/CASEID.fusioncatcher.fusion-genes.txt",
+        "path_index": "~",
+        "step": "fusioncatcher",
+        "tag": "fusioncatcher",
+    }
