+++ tests/models/nextflow/test_nextflow_deliver.py
@@ -1,93 +1,40 @@
 import pytest
-from cg.models.rnafusion.rnafusion_sample import RnafusionSample
+from cg.models.nextflow.deliverable import NextflowDeliverable
 import pydantic
 
 
-def test_instantiate_rnafusion_sample(
-    rnafusion_sample: str,
-    rnafusion_fastq_r1: list,
-    rnafusion_fastq_r2_same_length: list,
-    rnafusion_strandedness_acceptable: str,
+def test_instantiate_nextflow_deliverables(
+    nextflow_deliverables: dict,
 ):
     """
-    Tests rnafusion sample
+    Tests nextflow delivery object
     """
     # GIVEN a sample with fastq files and strandedness
 
-    # WHEN instantiating a MipAnalysis object
-    rnafusion_sample_object = RnafusionSample(
-        sample=rnafusion_sample,
-        fastq_r1=rnafusion_fastq_r1,
-        fastq_r2=rnafusion_fastq_r2_same_length,
-        strandedness=rnafusion_strandedness_acceptable,
-    )
+    # WHEN instantiating a deliverables object
+    nextflow_deliverables_object = NextflowDeliverable(deliverables=nextflow_deliverables)
 
     # THEN assert that it was successfully created
-    assert isinstance(rnafusion_sample_object, RnafusionSample)
+    assert isinstance(nextflow_deliverables_object, NextflowDeliverable)
 
 
-def test_instantiate_rnafusion_sample_fastq_r1_r2_different_length(
-    rnafusion_sample: str,
-    rnafusion_fastq_r1: list,
-    rnafusion_fastq_r2_not_same_length: list,
-    rnafusion_strandedness_acceptable: str,
+def test_instantiate_nextflow_deliverables_with_empty_entry(
+    nextflow_deliverables_with_empty_entry: dict,
 ):
     """
-    Tests rnafusion sample with different fastq_r1 and fastq_r2 length
+    Tests nextflow delivery object with empty entry
     """
-
-    # GIVEN a sample with fastq files and strandedness
-
-    # WHEN instantiating a sample object THEN throws a ValidationError
-
+    # WHEN instantiating a deliverables object with an empty entry THEN assert that it was successfully created
     with pytest.raises(pydantic.ValidationError):
-        RnafusionSample(
-            sample=rnafusion_sample,
-            fastq_r1=rnafusion_fastq_r1,
-            fastq_r2=rnafusion_fastq_r2_not_same_length,
-            strandedness=rnafusion_strandedness_acceptable,
-        )
+        NextflowDeliverable(deliverables=nextflow_deliverables_with_empty_entry)
 
 
-def test_instantiate_rnafusion_sample_fastq_r2_empty(
-    rnafusion_sample: str,
-    rnafusion_fastq_r1: list,
-    rnafusion_fastq_r2_empty: list,
-    rnafusion_strandedness_acceptable: str,
+def test_instantiate_nextflow_deliverables_with_faulty_entry(
+    nextflow_deliverables_with_faulty_entry: dict,
 ):
     """
-    Tests rnafusion sample with fastq_r2 empty (single end)
+    Tests nextflow delivery object with empty entry
     """
-
-    # GIVEN a sample with fastq files and strandedness
-
-    # WHEN instantiating a sample object
-    rnafusion_sample_object = RnafusionSample(
-        sample=rnafusion_sample,
-        fastq_r1=rnafusion_fastq_r1,
-        fastq_r2=rnafusion_fastq_r2_empty,
-        strandedness=rnafusion_strandedness_acceptable,
-    )
-
-    # THEN assert that it was successfully created
-    assert isinstance(rnafusion_sample_object, RnafusionSample)
-
-
-def test_instantiate_rnafusion_strandedness_not_acceptable(
-    rnafusion_sample: str,
-    rnafusion_fastq_r1: list,
-    rnafusion_fastq_r2_same_length: list,
-    rnafusion_strandedness_not_acceptable: str,
-):
-    """
-    Tests rnafusion sample with unacceptable strandedness
-    """
-    # WHEN instantiating a sample object THEN throws a ValidationError
-
+    # WHEN instantiating a deliverables object with an empty entry THEN assert that it was successfully created
     with pytest.raises(pydantic.ValidationError):
-        RnafusionSample(
-            sample=rnafusion_sample,
-            fastq_r1=rnafusion_fastq_r1,
-            fastq_r2=rnafusion_fastq_r2_same_length,
-            strandedness=rnafusion_strandedness_not_acceptable,
-        )
+        NextflowDeliverable(deliverables=nextflow_deliverables_with_faulty_entry)
