+++ tests/models/rnafusion/test_rnafusion_sample.py
@@ -1,40 +1,93 @@
-from cgmodels.cg.constants import Pipeline
-
-from cg.meta.workflow.mip import MipAnalysisAPI
-from cg.models.mip.mip_analysis import MipAnalysis
-
-
-def test_instantiate_mip_analysis(mip_analysis_raw: dict):
-    """
-    Tests raw mip analysis against a pydantic MipAnalysis
-    """
-    # GIVEN a dictionary with the some metrics
-
-    # WHEN instantiating a MipAnalysis object
-    mip_dna_analysis_object = MipAnalysis(**mip_analysis_raw)
-
-    # THEN assert that it was successfully created
-    assert isinstance(mip_dna_analysis_object, MipAnalysis)
+import pytest
+from cg.models.rnafusion.rnafusion_sample import RnafusionSample
+import pydantic
 
 
-def test_instantiate_parse_mip_analysis(
-    cg_context,
-    mip_analysis_config_dna_raw: dict,
-    mip_metrics_deliverables_raw: dict,
-    sample_info_dna_raw: dict,
+def test_instantiate_rnafusion_sample(
+    rnafusion_sample: str,
+    rnafusion_fastq_r1: list,
+    rnafusion_fastq_r2_same_length: list,
+    rnafusion_strandedness_acceptable: str,
 ):
     """
-    Tests parse_analysis
+    Tests rnafusion sample
     """
-    # GIVEN a dictionary with some metrics and a MIP analysis API
-    mip_analysis_api = MipAnalysisAPI(cg_context, Pipeline.MIP_DNA)
+    # GIVEN a sample with fastq files and strandedness
 
     # WHEN instantiating a MipAnalysis object
-    mip_dna_analysis_object = mip_analysis_api.parse_analysis(
-        config_raw=mip_analysis_config_dna_raw,
-        qc_metrics_raw=mip_metrics_deliverables_raw,
-        sample_info_raw=sample_info_dna_raw,
+    rnafusion_sample_object = RnafusionSample(
+        sample=rnafusion_sample,
+        fastq_r1=rnafusion_fastq_r1,
+        fastq_r2=rnafusion_fastq_r2_same_length,
+        strandedness=rnafusion_strandedness_acceptable,
     )
 
     # THEN assert that it was successfully created
-    assert isinstance(mip_dna_analysis_object, MipAnalysis)
+    assert isinstance(rnafusion_sample_object, RnafusionSample)
+
+
+def test_instantiate_rnafusion_sample_fastq_r1_r2_different_length(
+    rnafusion_sample: str,
+    rnafusion_fastq_r1: list,
+    rnafusion_fastq_r2_not_same_length: list,
+    rnafusion_strandedness_acceptable: str,
+):
+    """
+    Tests rnafusion sample with different fastq_r1 and fastq_r2 length
+    """
+
+    # GIVEN a sample with fastq files and strandedness
+
+    # WHEN instantiating a sample object THEN throws a ValidationError
+
+    with pytest.raises(pydantic.ValidationError):
+        RnafusionSample(
+            sample=rnafusion_sample,
+            fastq_r1=rnafusion_fastq_r1,
+            fastq_r2=rnafusion_fastq_r2_not_same_length,
+            strandedness=rnafusion_strandedness_acceptable,
+        )
+
+
+def test_instantiate_rnafusion_sample_fastq_r2_empty(
+    rnafusion_sample: str,
+    rnafusion_fastq_r1: list,
+    rnafusion_fastq_r2_empty: list,
+    rnafusion_strandedness_acceptable: str,
+):
+    """
+    Tests rnafusion sample with fastq_r2 empty (single end)
+    """
+
+    # GIVEN a sample with fastq files and strandedness
+
+    # WHEN instantiating a sample object
+    rnafusion_sample_object = RnafusionSample(
+        sample=rnafusion_sample,
+        fastq_r1=rnafusion_fastq_r1,
+        fastq_r2=rnafusion_fastq_r2_empty,
+        strandedness=rnafusion_strandedness_acceptable,
+    )
+
+    # THEN assert that it was successfully created
+    assert isinstance(rnafusion_sample_object, RnafusionSample)
+
+
+def test_instantiate_rnafusion_strandedness_not_acceptable(
+    rnafusion_sample: str,
+    rnafusion_fastq_r1: list,
+    rnafusion_fastq_r2_same_length: list,
+    rnafusion_strandedness_not_acceptable: str,
+):
+    """
+    Tests rnafusion sample with unacceptable strandedness
+    """
+    # WHEN instantiating a sample object THEN throws a ValidationError
+
+    with pytest.raises(pydantic.ValidationError):
+        RnafusionSample(
+            sample=rnafusion_sample,
+            fastq_r1=rnafusion_fastq_r1,
+            fastq_r2=rnafusion_fastq_r2_same_length,
+            strandedness=rnafusion_strandedness_not_acceptable,
+        )
